+++ src/components/gameslist/IdlingGamesList.jsx
@@ -1,72 +1,73 @@
+import { addToast, Button } from '@heroui/react';
 import { invoke } from '@tauri-apps/api/core';
-import Image from 'next/image';
 import { useContext } from 'react';
 import { TbPlayerStopFilled } from 'react-icons/tb';
 
 import { IdleContext } from '@/components/contexts/IdleContext';
+import GameCard from '@/components/ui/GameCard';
+import { logEvent } from '@/utils/utils';
 
 export default function IdlingGamesList() {
     const { idleGamesList, setIdleGamesList } = useContext(IdleContext);
 
-    const handleStopIdle = async (pid) => {
-        const response = await invoke('kill_process_by_pid', { pid });
-        if (response.includes('success')) {
-            setIdleGamesList(idleGamesList.filter((item) => item.pid !== pid));
+    const handleStopIdleAll = async () => {
+        try {
+            const pids = idleGamesList.map((item) => item.pid);
+            const response = await invoke('kill_all_process_by_pid', { pids });
+            if (response.includes('success')) {
+                setIdleGamesList([]);
+            } else {
+                addToast({ description: 'Failed to stop idling all games', color: 'danger' });
+            }
+        } catch (error) {
+            addToast({ description: 'Failed to stop idling all games', color: 'danger' });
+            console.error('Error in handleStopIdleAll:', error);
+            logEvent(`Error in (handleStopIdleAll): ${error}`);
         }
     };
 
     return (
-        <div className='w-calc min-h-calc max-h-calc bg-base overflow-y-auto overflow-x-hidden rounded-tl-xl border-t border-l border-border'>
-            <div className='fixed w-[calc(100vw-68px)] z-[50] bg-opacity-90 backdrop-blur-md bg-base pl-4 pt-2 pr-2 rounded-tl-xl'>
+        <div className='w-calc min-h-calc max-h-calc bg-base overflow-y-auto overflow-x-hidden rounded-tl-xl border-t border-l border-border select-none'>
+            <div className={`fixed w-[calc(100vw-68px)] z-[50] bg-opacity-90 backdrop-blur-md bg-base pl-4 pt-2 rounded-tl-xl ${idleGamesList?.length >= 21 ? 'pr-4' : 'pr-2'}`}>
                 <div className='flex justify-between items-center pb-3'>
-                    <div className='flex items-center gap-1'>
+                    <div className='flex items-center justify-between w-full gap-1'>
                         <div className='flex flex-col justify-center'>
                             <p className='text-lg font-semibold'>
                                 Idling Games
-                                <span className='text-[9px] text-altwhite border border-altwhite rounded-full px-1 ml-1 align-top'>
+                                <span className='text-[9px] text-altwhite border border-altwhite rounded-full px-1 ml-1 align-top select-none'>
                                     beta
                                 </span>
                             </p>
                             <div className='flex gap-1'>
                                 <p className='text-xs text-altwhite'>
-                                    Games that you are currently idling will be shown here
+                                    Games that you are currently idling will appear here
                                 </p>
                             </div>
+                        </div>
+                        <div>
+                            <Button
+                                size='sm'
+                                color='danger'
+                                isDisabled={idleGamesList?.length === 0}
+                                className='rounded-full font-semibold'
+                                startContent={<TbPlayerStopFilled fontSize={20} />}
+                                onPress={handleStopIdleAll}
+                            >
+                                Stop All
+                            </Button>
                         </div>
                     </div>
                 </div>
             </div>
 
-            <div className='p-4 pt-2'>
-                <div className='mt-[60px]'>
-                    <div className='grid grid-cols-5 2xl:grid-cols-7 gap-4'>
-                        {idleGamesList && idleGamesList.map((item) => (
-                            <div key={item.appid} className='relative group'>
-                                <div className='aspect-[460/215] rounded-xl overflow-hidden transition-transform duration-200 ease-in-out transform group-hover:scale-105'>
-                                    <Image
-                                        src={`https://cdn.cloudflare.steamstatic.com/steam/apps/${item.appid}/header.jpg`}
-                                        width={460}
-                                        height={215}
-                                        alt={`${item.name} image`}
-                                        priority={true}
-                                    />
-                                    <div className='absolute flex items-center justify-evenly inset-0 bg-black bg-opacity-0 dark:bg-opacity-20 group-hover:bg-opacity-40 dark:group-hover:bg-opacity-50 transition-opacity duration-200'>
-                                        <div className='absolute flex justify-center w-full bottom-0 left-0 px-2 pb-0.5 opacity-0 group-hover:opacity-100 duration-200'>
-                                            <p className='text-sm text-offwhite bg-black bg-opacity-50 rounded-sm px-1 select-none truncate'>
-                                                {item.name}
-                                            </p>
-                                        </div>
-                                        <div className='flex flex-col justify-center items-center opacity-0 group-hover:opacity-100 hover:scale-105 duration-200' onClick={() => handleStopIdle(item.pid)}>
-                                            <div className='p-2 bg-black text-offwhite bg-opacity-50 hover:bg-black hover:bg-opacity-70 cursor-pointer rounded-lg duration-200'>
-                                                <TbPlayerStopFilled className='text-offwhite opacity-0 group-hover:opacity-100 duration-200' fontSize={36} />
-                                            </div>
-                                        </div>
-                                    </div>
-                                </div>
-                            </div>
-                        ))}
-                    </div>
-                </div>
+            <div className='grid grid-cols-5 2xl:grid-cols-7 gap-4 p-4 pt-2 mt-[60px]'>
+                {idleGamesList && idleGamesList.map((item) => (
+                    <GameCard
+                        key={item.appid}
+                        item={item}
+                        setSettingsModalOpen={null}
+                    />
+                ))}
             </div>
         </div>
     );