+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 - Added `idling games` to the `sidebar` *(beta)*
   - This will hopefully replace the need for showing external SteamUtility *(idle)* windows for every game you idle
   - When a game begins idling it will be shown in the `idling games` list
-  - You can stop the game from idling by hovering over the `game card` and clicking the stop icon
+  - You can stop idling individual games by hovering over their `game card` and clicking the stop icon
+  - You can stop idling all games by clicking the `stop all` button
   - While this feature is in beta you will continue to see the external idle windows *(unless you use the `hide idle windows` setting)* just in case there are any problems
   - If you do have any problems, please [open a new issue](https://github.com/zevnda/steam-game-idler/issues/new?template=issue_report.yml)
+- #60: Added a pulsing border around `game cards` of currently idling games
 - #134: Fix `achievement unlocker` getting stuck at 1 second